Fixed negative ids that cause ArrayOutOfBounds.
When there are more than 32k resources with same type and package, tests crash with something like that:
java.lang.ArrayIndexOutOfBoundsException: Index -32362 out of bounds for length 34479
Commit fixes signed short overflow in this scenario.

+++ resources/src/main/java/org/robolectric/res/android/ResourceUtils.java
@@ -28,7 +28,7 @@ class ResourceUtils {
 
   static int get_entry_id(int resid) {
     //    return static_cast<uint16_t>(resid & 0x0000ffff);
-    return (short) (resid & 0x0000FFFF);
+    return resid & 0x0000FFFF;
   }
 
   static boolean is_internal_resid(int resid) {
